@@ -4,16 +4,18 @@
 	"description": "Access the system clipboard (copy/paste)",
 	"license": "MIT",
 	"repository": "sindresorhus/clipboard-cli",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
 	"bin": {
-		"clipboard": "cli.js"
+		"clipboard": "./cli.js"
 	},
 	"engines": {
-		"node": ">=8"
+		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
 	},
 	"scripts": {
 		"test": "xo && ava"
@@ -37,13 +39,13 @@
 		"xsel"
 	],
 	"dependencies": {
-		"clipboardy": "^2.0.0",
-		"get-stdin": "^7.0.0",
-		"meow": "^5.0.0"
+		"clipboardy": "^3.0.0",
+		"get-stdin": "^9.0.0",
+		"meow": "^10.1.1"
 	},
 	"devDependencies": {
-		"ava": "^1.4.1",
-		"execa": "^1.0.0",
-		"xo": "^0.24.0"
+		"ava": "^3.15.0",
+		"execa": "^5.1.1",
+		"xo": "^0.45.0"
 	}
 }
